refactor: Extract MapComponent logic to a separate non-SSR module
Why:
- Improved SSR support by dynamically loading MapComponent without server-side rendering.
- Enhanced maintainability by separating component logic into distinct files.
